First working prototype
The launch servo is working as expected based on two pin change interrupts.

The wing flapping servo is working as expected at the same time as the launch servo.

The flap servo function needs to be enabled/disabled using an external pin input for the final demo (still to be implemented)